+++ test/host-shim/atn.h
@@ -20,15 +20,15 @@ extern "C"
 #ifdef _SG3
 		#include "vartools.h"
 		#include "stringext.h"
-#endif
+#endif
 #ifdef _SG4
 		#include "vartools.h"
 		#include "stringext.h"
-#endif
+#endif
 #ifdef _SGC
 		#include "vartools.h"
 		#include "stringext.h"
-#endif
+#endif
 
 
 /* Datatypes and datatypes of function blocks */
@@ -58,6 +58,17 @@ typedef enum ATN_RESPONSE_ST
 	ATN_RESPONSE_ST_STATE_DONE,
 	ATN_RESPONSE_ST_BUSY
 } ATN_RESPONSE_ST;
+
+typedef enum AtnDiagSeverity_enum
+{	ATN_DIAG_INFO,
+	ATN_DIAG_WARNING,
+	ATN_DIAG_ERROR
+} AtnDiagSeverity_enum;
+
+typedef enum AtnDiagCode_enum
+{	ATN_DIAG_CODE_NONE = 0,
+	ATN_DIAG_CODE_PARAM_SIZE_MISMATCH = 1
+} AtnDiagCode_enum;
 
 typedef enum ATN_PLCOPEN_FUB_STATE_enum
 {	ATN_PLCOPEN_FUB_IDLE,
@@ -268,6 +279,9 @@ _BUR_PUBLIC plcbit stateStatus(plcstring* state, unsigned long buffer, unsigned 
 _BUR_PUBLIC plcbit systemJson(unsigned long buffer, unsigned long sBuffer);
 _BUR_PUBLIC plcbit resourceIsAvailable(plcstring* resourceName, unsigned long resourceUserId);
 _BUR_PUBLIC plcbit isInhibited(plcstring* inhibit);
+_BUR_PUBLIC signed long atnRaise(enum AtnDiagSeverity_enum severity, unsigned short code, plcstring* source, plcstring* message);
+_BUR_PUBLIC unsigned long atnDiagnosticCount(void);
+_BUR_PUBLIC signed long atnSetDiagnosticLogger(plcstring* loggerName);
 
 
 /* Constants */